requirements.txt
# 目前项目只使用 Python 标准库
# 未来可能会添加的依赖：
# openai>=1.0.0
# python-dotenv>=0.19.0

# 核心依赖
aiofiles==23.2.1
GitPython==3.1.42
openai==1.12.0
python-dotenv==1.0.1

# 开发依赖
pytest==8.0.2     # 单元测试
pytest-asyncio==0.23.5  # 异步测试支持

# 未来可能添加的依赖
# openai>=1.0.0
# python-dotenv>=0.19.0 